# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Cottage Village API Documentation",
        "description": "",
        "contact": {
            "email": "afonico@info.ru"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/profile/change/password": {
            "post": {
                "tags": [
                    "User"
                ],
                "summary": "Change auth users password",
                "description": "",
                "operationId": "changePassword",
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "password",
                                    "new_password",
                                    "new_password_confirmation"
                                ],
                                "properties": {
                                    "password": {
                                        "description": "Current users password",
                                        "type": "string"
                                    },
                                    "new_password": {
                                        "description": "New users password",
                                        "type": "string"
                                    },
                                    "new_password_confirmation": {
                                        "description": "New users password confirmation",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Ok"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
        "/api/profile/change/email": {
            "post": {
                "tags": [
                    "User"
                ],
                "summary": "Change auth users email",
                "description": "",
                "operationId": "changeEmail",
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "email",
                                    "new_email"
                                ],
                                "properties": {
                                    "email": {
                                        "description": "Current users email",
                                        "type": "string"
                                    },
                                    "new_email": {
                                        "description": "New users email",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Ok"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
        "/api/profile/change": {
            "post": {
                "tags": [
                    "User"
                ],
                "summary": "Change auth users profile data",
                "description": "",
                "operationId": "changeProfile",
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "description": "Users name. Size should be less than 100 characters",
                                        "type": "string"
                                    },
                                    "surname": {
                                        "description": "Users surname. Size should be less than 100 characters",
                                        "type": "string"
                                    },
                                    "birthday": {
                                        "description": "Users birthday date",
                                        "type": "string",
                                        "format": "date"
                                    },
                                    "gender_id": {
                                        "description": "User gender_id, 1 = Male, 2 = Female",
                                        "type": "integer",
                                        "default": 1,
                                        "enum": [
                                            1,
                                            2
                                        ]
                                    },
                                    "phone": {
                                        "description": "Users phone. Unique value, should be masked value in format +7 (111) 222-33-44, server proceed to format +71112223344 to save in DB and returns to client in masked format",
                                        "type": "string"
                                    },
                                    "whatsapp_phone": {
                                        "description": "Should be masked value in format +7 (111) 222-33-44",
                                        "type": "string"
                                    },
                                    "telegram_phone": {
                                        "description": "Should be masked value in format +7 (111) 222-33-44",
                                        "type": "string"
                                    },
                                    "viber_phone": {
                                        "description": "Should be masked value in format +7 (111) 222-33-44",
                                        "type": "string"
                                    },
                                    "work_info": {
                                        "description": "Size should be less than 1000 characters. Can be empty",
                                        "type": "string"
                                    },
                                    "hobby_info": {
                                        "description": "Size should be less than 1000 characters. Can be empty",
                                        "type": "string"
                                    },
                                    "family_info": {
                                        "description": "Size should be less than 1000 characters. Can be empty",
                                        "type": "string"
                                    },
                                    "extra_info": {
                                        "description": "Size should be less than 1000 characters. Can be empty",
                                        "type": "string"
                                    },
                                    "user_estates": {
                                        "description": "User estates data. estate_type_id variations is { 1: 'Участок', 2: 'Квартира'}. estate_number should be greater or equal 0. estate_number required when picked estate_type_id. If you want to update estate you should pass 'id' parameter",
                                        "type": "string",
                                        "format": "array",
                                        "example": "{ 'user_estates': [ { 'id': 1, 'estate_type_id': 1, estate_number: 10 } ] }"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Ok"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
...
    },
...
    "components": {
        "securitySchemes": {
            "bearer_token": {
                "type": "apiKey",
                "description": "Enter token in format: token_type access_token",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
    "security": [
        []
    ]
}